// src/interfaces/Interfaces.ts
export interface IKalturaConfigurations {
  kalturaAdminSecret: string;
  kalturaUserSecret: string;
  kalturaPartnerId: number;
  kalturaSubPartnerId: number;
  kalturaUiConfigId: number;
  kalturaAppHost: string;
  kalturaCuePointUrl: string;
  kalturaServiceUrl: string;
  kalturaEndPointUrl: string;
  kalturaPrivilegesForSelfCapture: string;
  kalturaPrivilegesForOptions: string;
  kalturaVideoEditorSource: string;
  kalturaStatsSource: string;
  kalturaPlayerSourceUrl: string;
  kalturaVideoOptionsUrl: string;
}